RFC: default value hash
Summary: I went down the road we discussed and implemented
two different hashes for config types: a "structural" hash
in addition to the key. I was quite dissatisifed by the complexity
and conceptual confusion this incurred.

As another alternative approach, where instead of using
a raw str cast we instead serialize the config value using
dagster.serdes and then hash that result.

This results in a stable hash across python versions (so far). However
it requires whitelisting everything. In our test cases, that only
results in having to annotate a single Enum. However this
might be an unacceptable burden for users, because there are
mitigations to discuss.

Depends on D2361

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D2362

Reviewers: alangenfeld, themissinghlink